requirements.txt
frida
frida-tools
protobuf
pycryptodome
